moved cookie sign in to context
