+++ state/appContext.js
@@ -1,5 +1,41 @@
-import { createContext, useContext, useState } from 'react';
+import { createContext, useContext, useState, useEffect } from 'react';
+import { gql, useLazyQuery } from '@apollo/client';
+import cookie from 'cookie';
 import { useToggle } from 'hooks';
+
+const REFRESH_USER = gql`
+  query getUser($id: ID) {
+    userById(id: $id) {
+      id
+      username
+      name {
+        first
+        last
+      }
+      fullname
+      events {
+        id
+        name
+        admin {
+          id
+          username
+        }
+        participants {
+          id
+          username
+        }
+        active
+        private
+        rounds
+        shotsPer
+        scores
+        participantCap
+        startDate
+        endDate
+      }
+    } 
+  }
+`;
 
 const AppContext = createContext({
   isLoginOpen: false,
@@ -11,6 +47,17 @@ export const AppContextWrapper = ({ children }) => {
   const [isLoginOpen, setIsLoginOpen, toggleLogin] = useToggle(false, true);
   const [isSignUpOpen, setIsSignUpOpen, toggleSignUp] = useToggle(false, true);
   const [toasts, setToasts] = useState([]);
+  const [getUser, { data, loading: userLoading, error: userLoadError }] = useLazyQuery(REFRESH_USER, {
+    onError: err => console.log(err),
+    onCompleted: () => setUser(data.userById)
+  });
+  useEffect(() => {
+    const COOKIES = cookie.parse(document.cookie);
+    if (!user && COOKIES.LQ_USER) {
+      getUser({ variables: { id: COOKIES.LQ_USER } });
+
+    }
+  }, []);
 
   return (
     <AppContext.Provider
@@ -34,7 +81,9 @@ export const AppContextWrapper = ({ children }) => {
         removeToast: id => {
           const newList = toasts.filter(toast => toast.id !== id);
           setToasts(newList);
-        }
+        },
+        userLoading,
+        userLoadError,
       }}
     >
       {children}
